Add PerformanceLogger Interface to PerL
